feat: accept async iterators into blockstore.putMany
Allows streaming into the blockstore. The blockstore currently uses
the batch functionality of the underlying datatore, so we might need
to change `interface-datastore` next to add streaming primitives but
this at least allows streaming this far down the stack.

BREAKING CHANGE: you must pass an iterable or async iterable to putMany
- this should be relatively painless as the current API is to pass an
array which is iterable, but it does change the API.