feat: add isInitialized method
A few places need ot know whether or not a repo has been initialized
and do so by a variety of methods - checking whether certain files
exist, etc.

This PR adds an `isInitialized()` method to the ipfs repo class to
take some of this guesswork away as the repo will know if it's been
initialized.