feat: adds .replace and .getAll methods to config (#227)
The config .get/.set methods are overloaded to allow the user to retrieve or replace the entire config.  This is an error prone approach that could end with the user invalidating their config when they did not mean to.

The change here adds `.replace` and `.getAll` methods to allow the user to be specific about their intentions and hopefully makes the API slightly less of a foot-gun.